Decouple `TestingFarmClient` from `TestingFarmJobHelper`
